// src/enums/ageRange.ts
export enum AgeRange {
    THIRTEEN_TO_EIGHTEEN = 1318,
    NINETEEN_TO_TWENTY_FOUR = 1924,
    TWENTY_FIVE_TO_THIRTY = 2530,
    THIRTY_ONE_TO_THIRTY_SIX = 3136,
    THIRTY_SEVEN_TO_FORTY_TWO = 3742,
    FORTY_THREE_TO_FORTY_EIGHT = 4348,
    FORTY_NINE_TO_FIFTY_FOUR = 4954,
    FIFTY_FIVE_TO_SIXTY = 5560
}