// src/enums/weightRange.ts
export enum WeightRange {
    FORTY_TO_FIFTY = 4050,
    FIFTY_ONE_TO_SIXTY = 5160,
    SIXTY_ONE_TO_SEVENTY = 6170,
    SEVENTY_ONE_TO_EIGHTY = 7180,
    EIGHTY_ONE_TO_NINETY = 8190,
    NINETY_ONE_TO_HUNDRED = 91100,
    HUNDRED_ONE_TO_ONE_TEN = 101110,
    ONE_TEN_ONE_TO_ONE_TWENTY = 111120,
    ONE_TWENTY_ONE_TO_ONE_THIRTY = 121130,
    ONE_THIRTY_ONE_TO_ONE_FORTY = 131140,
    ONE_FORTY_ONE_TO_ONE_FIFTY = 141150,
    ONE_FIFTY_ONE_TO_ONE_SIXTY = 151160,
    ONE_SIXTY_ONE_TO_ONE_SEVENTY = 161170,
    ONE_SEVENTY_ONE_TO_ONE_EIGHTY = 171180,
    ONE_EIGHTY_ONE_TO_ONE_NINETY = 181190,
    ONE_NINETY_ONE_TO_TWO_HUNDRED = 191200
}