// src/enums/plankDurationRange.ts
export enum PlankDurationRange {
    FIFTEEN_TO_FORTY_FIVE = 1545,
    FORTY_SIX_TO_SEVENTY_FIVE = 4675,
    SEVENTY_SIX_TO_ONE_HUNDRED_FIVE = 76105,
    ONE_HUNDRED_SIX_TO_ONE_THIRTY_FIVE = 106135,
    ONE_THIRTY_SIX_TO_ONE_SIXTY_FIVE = 136165,
    ONE_SIXTY_SIX_TO_ONE_NINETY_FIVE = 166195,
    ONE_NINETY_SIX_TO_TWO_TWENTY_FIVE = 196225,
    TWO_TWENTY_SIX_TO_TWO_FIFTY_FIVE = 226255,
    TWO_FIFTY_SIX_TO_TWO_EIGHTY_FIVE = 256285,
    TWO_EIGHTY_SIX_TO_THREE_FIFTEEN = 286315,
    THREE_SIXTEEN_TO_THREE_FORTY_FIVE = 316345,
    THREE_FORTY_SIX_TO_THREE_SEVENTY_FIVE = 346375,
    THREE_SEVENTY_SIX_TO_FOUR_HUNDRED_FIVE = 376405,
    FOUR_HUNDRED_SIX_TO_FOUR_THIRTY_FIVE = 406435,
    FOUR_THIRTY_SIX_TO_FOUR_SIXTY_FIVE = 436465,
    FOUR_SIXTY_SIX_TO_FOUR_NINETY_FIVE = 466495,
    FOUR_NINETY_SIX_TO_FIVE_TWENTY_FIVE = 496525,
    FIVE_TWENTY_SIX_TO_FIVE_FIFTY_FIVE = 526555,
    FIVE_FIFTY_SIX_TO_FIVE_EIGHTY_FIVE = 556585,
    FIVE_EIGHTY_SIX_TO_SIX_HUNDRED = 586600
}